+++ examples/advanced/spline_draw.py
@@ -4,7 +4,7 @@ from vedo.applications import SplinePlotter  # ready to use class!
 pic = Picture(dataurl + "images/embryo.jpg")
 
 plt = SplinePlotter(pic)
-plt.show(mode="image", zoom='tight')
+plt.show(mode="image", zoom='tightest')
 print("Npts =", len(plt.cpoints), "NSpline =", plt.line.npoints)