+++ examples/volumetric/slice_plane1.py
@@ -12,7 +12,7 @@ def func(evt):
     sph = Sphere(pts[pid], c='orange7').pickable(False)
     fp = sph.flagpole(txt, s=7, offset=(-150,15), font=2).follow_camera()
     # remove old and add the two new objects
-    plt.remove('Sphere', 'FlagPole').add(sph, fp)
+    plt.remove('Sphere', 'FlagPole').add(sph, fp).render()
 
 vol = Volume(dataurl+'embryo.slc').alpha([0,0,0.8]).c('w').pickable(False)
 vslice = vol.slice_plane(origin=vol.center(), normal=(0,1,1))